requirements.txt
asgiref==3.5.0
certifi==2021.10.8
Django==4.0.4
djangorestframework==3.13.1
mysqlclient==2.1.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
